FEATURE: Validate Mentions in Rich Text Editor (#32879)
This change validates mentions for `users` and `groups` when using the
rich text editor.

When the mention is valid (ie. the user or group exists) it will
visually show as a mention node, but when the user or group does not
exist then it will appear as regular text in the editor.

There are 2 parts to this:

### Existing drafts / toggling from Markdown to Rich Text

We automatically add mention nodes for all `@mentions` and then process
any invalid mentions by removing the mention node and replacing it with
text.

### When manually typing in Rich Text

We find the mention within the text and then if valid we replace it with
a mention node.

### Validation

The validation happens in a single request when loading a draft or
switching from Markdown to Rich Text. The response provides a list of
valid usernames and groups. It also provides additional context for
groups and users that are not reachable (ie. unmentionable groups or
suspended users etc) but we don't use this currently.

We then store the valid and invalid mentions to prevent unnecessary
requests later.

As new mentions are typed then they will also be validated using the
same request as above (if they are not already stored in our valid /
invalid sets).