FIX: Allow setting blank secret on an existing webhook (#33096)
We had a regression where you can no longer update an existing webhook to have a blank secret field.

The problem is the blank field will return a value of undefined, which is ignored by the store when sending the request to the back-end.

This fixes that by using the nullish coalescing operator to replace undefined with an empty string.