FEATURE: sharing link to PWA renders url in title in new Topic to leverage core link expansion (#32631)
Discourse has a fantastic feature whereby when installed as PWA (on
Android or Windows), you can share content to that Discourse from other
apps and browsers.

Currently when sharing a link to a Discourse PWA, it pre-populates a
draft Topic body with the link.

Whilst this is really nice already, we can improve this further.
Currently, sharing a link to the body misses out on the auto-magic core
feature of pasting a link in the title which simultaneously hydrates a
title from the og tag and creates a one-box in the body.

This PR remedies this whilst carefully maintaining the current behaviour
for text shares.

Documentation for the standard is here:


https://developer.mozilla.org/en-US/docs/Web/Progressive_web_apps/How_to/Share_data_between_apps

Demo of this working change is implemented in this plugin:

https://github.com/merefield/discourse-share-to-link-oneboxer

I added a test to check presence of share target in manifest endpoint.