DEV: Adds `hasNoPreferredMode` to chat state manager
This commit introduces a new property `hasNoPreferredMode` to the
chat state manager, which represents a user who has not purposely
set chat mode to drawer or full page, meaning they have no LocalStorage
value set.

This can be useful for themes to change the chat mode but only
if the user has no preference already.

c.f. https://meta.discourse.org/t/full-screen-chat-as-default-for-collaboration-setup/369849