FIX: Multiple video uploads in composer failed
When uploading > 1 video in the composer, any upload after
the first one would fail silently, and not show anything in
the composer.

This was happening because we were removing uploads before the video
thumbnail callback was fired, but that had code like this `if
(this.#inProgressUploads.length === 0) { this.#reset() }`, which meant
that before the second upload was completely done we were resetting the
state.

Instead, we can remove the in progress upload file inside the video
thumbnail callback, which will ensure that the state is not reset
until all uploads are done.