FIX: prevent login loop when logging out when only one idp
When we migrated to the full page /login, we added a nice feature that will automatically redirect a user to the idp if it's the only way they can log in, thus avoiding unecessary click.

But when logging out, we would redirect to the /login page which would start the /login process by automatically redirecting the user to the idp...

The fix is easy, redirect back to / after logging out instead. This will show the "splash" screen that asks the user to either log in or sign up.

Note: I also removed the `Discourse::Utils::EMPTY_KEYWORDS` since it was its last occurence and I'm pretty sure we dropped support for Ruby 2 a while ago...

Internal ref - t/156834