DEV: Bump aws-sdk-core in prep for aws-sdk-mediaconvert (#33250)
In an upcoming pr #33092 that adds support for aws-sdk-mediaconvert it
requires
that we bump aws-sdk-core to 3.225.0

```
+    aws-sdk-mediaconvert (1.160.0)
+      aws-sdk-core (~> 3, >= 3.225.0)
```

but in doing so there are apparently some api changes and it requires
changes
in core to support the new version. I thought it would best to break out
these
changes in a separate PR as to not muddy up the mediaconvert pr that is
already
quite large.