FIX: improvements to keyboard shortcuts from toolbar API (#33290)
## :mag: Overview

This update fixes a few issues related to keyboard shortcuts from the
toolbar API and the textarea text manipulation:
- it converts the keymap return value from `itsatrap` format to
`ProseMirror` format so that the default is prevented, thus allowing for
custom keyboard shortcuts to work inside the editor without triggering
system defaults.
- pushes a fix where the keyboard shortcut could get triggered even when
the `condition` value returned `false` whenever it had a custom
shortcut.
- fixes an issue in the markdown textarea editor where selecting text
and tabbing works but also focuses the next element