FIX: Return the post's raw when GET /post (#33325)
Previously, there was no way to edit post localizations, so the
`post.raw` was always returning the localized value. With localization
UI and endpoints now, we want to always return the original raw and the
composer to reflect the original value.

https://github.com/discourse/discourse/pull/33184 will handle the
decision to show the post edit composer or the post localization
composer.