DEV: Introduce AdminNavManager service for admin sidebar and search links
Followup d92de3e4d057a785187ea3923313a26f5241c6b2

We have an ADMIN_NAV_MAP const that was being used in the admin sidebar
and the admin search to generate a list of links that the admin can
see and navigate to.

However, we were modifying this const, which led to inconsistent
behaviour between the admin sidebar and admin search, leading to
flaky tests.

This commit moves the access control of ADMIN_NAV_MAP into a new
service, and introduces a more structured way of adding/changing
links in the nav map that the sidebar and search can use directly.

This commit also removes old unused code from the admin sidebar
state manager used to store a custom admin nav map in localstorage.