FIX: improves post toolbar behavior (#33342)
This commit is applying different techniques to make selecting text of a
post less error prone:
- ensure we only ever show the toolbar when the common ancestor of a
range is cooked
- ensures the menu is not interfering with text selection

The situation was very bad on android but it should also improve other
devices.

Note that one visual change of this commit is that the fast edit menu is
now a standalone modal.