UX: delete components without full reload to maintain pagination (#33364)
Reported here:
https://meta.discourse.org/t/components-list-not-showing-all-components-after-deleting-one/371784

Deleting a component reloads the component list, but doesn't maintain
pagination if the admin loaded multiple pages of components... so you
can end up with only a partial list after delete.

This instead removes the component from the array without a reload, so
it maintains the current list state.