DEV: Add various APIs and outlets to the themes config area (#33385)
This commit adds plugin outlets throughout the themes and components admin pages to allow plugins to insert and replace various elements on those pages, and adds a plugin API to append custom cards to the themes grid at `/admin/config/customize/themes`.

A refactor has been done to move the `AdminConfigAreaCard` component out of the `admin/` bundle into the main/app bundle to make the plugin API work without breaking the site for non-staff users. If the component file stayed in the admin bundle, the plugin API class, which is in the main bundle, would fail to import from the component file in the admin bundle the required functions when normal users accessed the site because the admin bundle wouldn't be loaded for them.

Internal topic: t/156924.