FIX: ensures small actions don't trigger post toolbar
Prior to this fix you could select the text of a small action (after closing a topic with a message for example) and it would show the toolbar.

This commit will now ensure the common ancestor of the text range has no ".small-action" node as parent.

No test as it's currently very hard to test this behavior due to the debouncing delay, I need to re-think few parts of the toolbar to make it reliable.