A11Y: make in-reply-to keyboard accessible (#33447)
Adding an empty href to this link makes it keyboard accessible. You can
now trigger it with the enter key. This also allows us to remove the
tabindex.


![image](https://github.com/user-attachments/assets/357f7a6e-05fd-46dc-9da7-56520e9ca8f0)


`event.preventDefault();` is neccessary because otherwise `href=""` was
resulting in a page refresh, and another work around like `href="#"` was
adding an entry to the browser history