FIX: Make hide_new_user_profiles work with manually upgraded users (#33458)
The site setting hide_new_user_profiles states:

> Hide trust level 1 or lower user profiles from the public and trust level 1 users until they post for the first time.

This implies that TL2 and above should be visible.

However, the actual logic encodes the assumption that TL2 have posted at some point. This fails to account for the fact that admins can manually upgrade a user without posts to a higher TL. This PR covers for that scenario.