FIX: clear selection on expand popup menu
Having a text selection can cause issues on iOS where the caret will show over any other element. The state is already stored in the toolbar state so we don't need to keep it actually selected.

No test as it's a very specific behavior and a whole system spec for this seems heavy.