FEATURE: rich editor image toolbar for scale/alt text/delete (#33381)
Displays a toolbar and the alt text of the image when it's selected.

The toolbar allows changing the scale between 50, 75 and 100 and
deleting the image.

The alt text display can be clicked/focused to open an editable input.



https://github.com/user-attachments/assets/330f5c7c-e29a-46ab-b2cd-e7c7c07c15bf



It also introduces a simple GlimmerNodeView class that serves as a
bridge between ProseMirror NodeViews and Glimmer components, which is
used on this new version of the image node.

---------

Co-authored-by: chapoi <101828855+chapoi@users.noreply.github.com>
Co-authored-by: Joffrey JAFFEUX <j.jaffeux@gmail.com>
Co-authored-by: Martin Brennan <martin@discourse.org>