FIX: convert invalid mentions in composer rich text mode to text (#33437)
This change makes it easier to edit typos when adding mentions to
composer's rich text mode. Previously we would keep invalid mentions as
mention nodes and use a data attribute to identify if they were valid or
not and style them based on that. This would mean that pressing
backspace would delete the entire mention and you would have to type
again from scratch.

The updated approach is to replace the invalid mention node with text
and in turn removes the need to use the data attribute to identify
valid/invalid mentions.