FIX: convert invalid hashtags in composer to text (#33441)
This change makes it easier to edit typos when adding hashtags to
composer's rich text mode. Previously we would keep invalid hashtags as
hashtag nodes and use a data attribute to identify if they were valid or
not and style them based on that. This would mean that pressing
backspace would delete the entire hashtag and you would have to type
again from scratch.

The updated approach is to replace the invalid hashtag node with text
and in turn removes the need to use the data attribute to identify
valid/invalid hashtags.