FEATURE: image input rule when typing on rich editor (#33498)
Adds a `![markdown](image)` input rule that uses the markdown parser to
create an image node.

Supporting this format surfaced an issue with dimension-less images, so
now the image node view uses the max width/height site settings in
addition to the loaded image natural dimensions to determine what a 100%
looks like – it's used for display, and also when a zoom-out button
action is taken, so we calculate the next step % of the calculated
dimensions.