FIX: Use the max_tag_search_results setting as the default limit for tag groups search (#33485)
Hardcoding the default limit and using a setting for the max value could
result in an error if the setting is changed to anything lower than the
hardcoded value for the default limit. It's better to use the setting
for the default value for the param and use the same max limit for the
setting as the max limit for the param.

Internal topic: t/156246.