UX: Variable additions (#33473)
Working on adding a lot more variables to core. Doing so enables easy
styling & changes to the main bulk of the interface.

**Example of what can be done**
Using these variables in core allows for the following stylistic changes
to be made with less than 100 lines of code. Most of which are changing
the value assigned to variables in this PR.

![CleanShot 2025-07-03 at 19 56
46@2x](https://github.com/user-attachments/assets/130dc8be-3970-49ea-9003-d124d779edcc)