DEV: Load admin JS bundles from plugins for staff users (stable) (#33556)
JS assets defined under `admin/` in plugins are compiled into a separate
bundle that should be loaded for staff users only, but we currently load
it for all users including normal users and anons. This is wasteful and
unexpected since it could cause errors if an asset in the admin bundle
of a plugin imports a module from the admin bundle in core (which
doesn't get loaded for non-staff users).

`main` PR: https://github.com/discourse/discourse/pull/33540.