UX: edit category description layout fix (#33565)
The `form-kit` container subtitle doesn't play well with the category
description. The issue experienced appears to be due to the flex
direction.

Since we can only update the category description from the topic and not
within the edit category page, I have moved the description to a
read-only span which follows a similar pattern that we use on other
forms.