DEV: Add X-Discourse-BrowserPageView response header
When we are tracking requests in the `Middleware::RequestTracker`, we
have historically added an `X-Discourse-TrackView` response header
for implicitly tracked requests, which are HTML requests.

Within the past couple of years, we introduced the concept of
Browser Page Views (BPVs), which are also known as explicitly tracked
requests, which are recorded when requests are made via AJAX.
These also add the `X-Discourse-TrackView` header to the response,
so it is hard to differentiate the two in logs.

This commit adds a new response header, `X-Discourse-BrowserPageView`,
that acts in a similar way to the existing `X-Discourse-TrackView`
header, but is specifically for requests that are tracked as BPVs.