DEV: Move post-related specs from guardian_spec to post_guardian_spec (#33587)
This is the first in a series of PRs aimed at organizing the guardians and their tests a bit better to make them more discoverable and reduce the amount of "pinballing" around looking for the relevant methods.

There is no change to any functionality, and the tests have been copied over as-is, except where fixture variables needed renaming to fit in the new spec file.