FIX: Broken links to site settings under Content section (#33597)
The Content section in particular has four tabs which are all site setting pages. It does not have one under /settings, like other sections.

The admin search functionality will automatically use /settings for multi-tabbed sections. This results in a 404 when searching for Content site settings.

This PR adds a front-end redirect (replaceWith) for /content/settings to fix the immediate issue.