DEV: Extract theme script tags into their own JS files
Previously, all script tags in a theme field would be combined into a single `.js` bundle along with the `text/discourse-plugin` modules. This led to some unexpected run order, and also makes it hard to implement the `type=module` change being developed in #33103.

This commit refactors things so that each raw script gets extracted into its own `.js` bundle, which is then placed in exactly the same place in the HTML.